+++ app/pages/userprofile/userprofile.js
@@ -74,6 +74,7 @@ export var UserProfile = translate()(React.createClass({
         items: [
           {value: 'en', label: 'English'},
           {value: 'fr', label: 'Français'},
+          {value: 'de', label: 'Deutsche'},
         ],
         placeholder: t('Select language...')
       });