+++ CHANGELOG.md
@@ -6,6 +6,7 @@ It is based on Tidepool Blip 1.27.
 ### Added
 - PT-401 Display rescueCarbs with specific layout in Daily view
 - PT-574 Integrate latest change from [Tidepool 1.27.0](https://github.com/tidepool-org/blip/releases/tag/v1.27.0)
+- PT-415 Add german language
 
 ## [0.6.2] - 2019-09-16
 - PT-640 Upgrade viz to dblp.0.4.3